Cleanup post https://github.com/finos/legend-studio/pull/1367
