+++ packages/legend-application-query/src/components/__tests__/TEST_DATA__QueryBuilder_Model_HiglightProperties.json
