+++ packages/legend-graph/src/graphManager/__tests__/roundtripTestData/executionPlan/TEST_DATA__SimpleRelationalPlan.ts
@@ -186,32 +186,11 @@ export const TEST_DATA__simpleRelationalPlan = {
                       {
                         _type: 'var',
                         name: 'this',
-                        sourceInformation: {
-                          endColumn: 23,
-                          endLine: 35,
-                          sourceId: '',
-                          startColumn: 19,
-                          startLine: 35,
-                        },
                       },
                     ],
                     property: 'employees',
-                    sourceInformation: {
-                      endColumn: 33,
-                      endLine: 35,
-                      sourceId: '',
-                      startColumn: 25,
-                      startLine: 35,
-                    },
                   },
                 ],
-                sourceInformation: {
-                  endColumn: 40,
-                  endLine: 35,
-                  sourceId: '',
-                  startColumn: 36,
-                  startLine: 35,
-                },
               },
             ],
             name: 'employeeSize',
